Unite common code.
